config yml libraryfixed

+++ libraryapi/src/main/java/io/github/educastrob/libraryapi/service/AutorService.java
@@ -1,7 +1,10 @@
 package io.github.educastrob.libraryapi.service;
 
+import io.github.educastrob.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.educastrob.libraryapi.model.Autor;
+import io.github.educastrob.libraryapi.model.Livro;
 import io.github.educastrob.libraryapi.repository.AutorRepository;
+import io.github.educastrob.libraryapi.repository.LivroRepository;
 import io.github.educastrob.libraryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
@@ -14,10 +17,12 @@ public class AutorService {
 
     private final AutorRepository repository;
     private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository, AutorValidator validator) {
+    public AutorService(AutorRepository repository, AutorValidator validator, LivroRepository livroRepository) {
         this.repository = repository;
         this.validator = validator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor) {
@@ -38,6 +43,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if (possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException("Autor possui livros cadastrados!");
+        }
         repository.delete(autor);
     }
 
@@ -55,5 +63,9 @@ public class AutorService {
         }
 
         return repository.findAll();
+    }
+
+    public boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
     }
 }
